+++ tests/test_to_thread.py
@@ -114,6 +114,22 @@ async def test_cancel_worker_thread(cancellable: bool, expected_last_active: str
     assert last_active == expected_last_active
 
 
+async def test_cancel_wait_on_thread() -> None:
+    def wait_event() -> None:
+        nonlocal event_triggered
+        event_triggered = event.wait(1)
+
+    event_triggered: bool = False
+    event = threading.Event()
+    async with create_task_group() as tg:
+        tg.start_soon(partial(to_thread.run_sync, cancellable=True), wait_event)
+        await wait_all_tasks_blocked()
+        tg.cancel_scope.cancel()
+
+    await to_thread.run_sync(event.set)
+    assert event_triggered
+
+
 @pytest.mark.parametrize('anyio_backend', ['asyncio'])
 async def test_asyncio_cancel_native_task() -> None:
     task: "Optional[asyncio.Task[None]]" = None